# docs/api-contract.yaml
components:
...
  schemas:
...
    AuthVerifyBody:
      additionalProperties: false
      properties:
        mode:
          $ref: '#/components/schemas/MagicLinkMode'
          examples:
          - signup
        token:
          examples:
          - ml-tok-abcdef0123456789
          minLength: 8
          title: Token
          type: string
      required:
      - token
      - mode
      title: AuthVerifyBody
      type: object
    AuthVerifyResponse:
      additionalProperties: false
      example:
        email: alice@example.com
        is_first_time: true
        pre_totp_token: pre-totp-abcdef0123456789
        requires_totp_challenge: false
        requires_totp_enrollment: true
        user_id: 11111111-1111-1111-1111-111111111111
      properties:
        email:
          examples:
          - alice@example.com
          maxLength: 254
          minLength: 3
          title: Email
          type: string
        is_first_time:
          examples:
          - true
          title: Is First Time
          type: boolean
        pre_totp_token:
          examples:
          - pre-totp-abcdef0123456789
          title: Pre Totp Token
          type: string
        requires_totp_challenge:
          examples:
          - false
          title: Requires Totp Challenge
          type: boolean
        requires_totp_enrollment:
          examples:
          - true
          title: Requires Totp Enrollment
          type: boolean
        user_id:
          examples:
          - 11111111-1111-1111-1111-111111111111
          format: uuid
          title: User Id
          type: string
      required:
      - user_id
      - email
      - is_first_time
      - requires_totp_enrollment
      - requires_totp_challenge
      - pre_totp_token
      title: AuthVerifyResponse
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    MagicLinkMode:
      description: 'Magic-link request kind. Same string values the brief calls out

        so wire-format and tests stay readable.'
      enum:
      - signup
      - login
      title: MagicLinkMode
      type: string
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  title: VaultChain API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/v1/auth/verify:
    post:
      description: 'Consume the raw magic-link token; on success mint a 5-min Redis-cached

        pre-TOTP bearer token that gates the TOTP routes.'
      operationId: auth_verify_api_v1_auth_verify_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthVerifyBody'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthVerifyResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Consume a magic link and mint a pre-TOTP token
      tags:
      - auth
...